fix: move types to a separate file; create City type
